Added dao_relational and util modul, configure database connection and create test class
Added dao_relational modul and configure mysql database connection by hibernate. In addition defined datas which got to database. It came to not handled enum type in Skill class. Unfortunately not gaven to database. Therefore Skill enum type changed String. So must be typed by hand the skills, name but I not deleted file. This will be able to help